+++ CHANGELOG.md
@@ -14,6 +14,7 @@ n/a
 * Fixed a bug where the `DarkmodeDialog` did not display the UI mode of the Android OS but instead used the UI mode of the app.
 * Changed color of background when swiping an item to delete from _@color/pv\_red_ to _@color/text\_critical_.
 * Implemented new framework for recycler item swiping.
+* Added `MoreDialog` which dynamically displays a list of clickable items to the user.
 
 <br/>